refactor(user): encapsulate delete account logic inside DangerZone for better modularity
