+++ src/views/UserDashboard/UserDashboard.tsx
@@ -1,5 +1,4 @@
 import { Box, Grid, Typography, Container } from "@mui/material";
-import { useNavigate } from "react-router-dom";
 import { useUserProfile } from "./useUserProfile";
 import { useScores } from "./ScoresSection/useScores";
 import AvatarSection from "./AvatarSection/AvatarSection";
@@ -9,37 +8,15 @@ import ChangePasswordForm from "./ChangePasswordForm/ChangePasswordForm";
 import UserDataInfo from "./UserDataInfo/UserDataInfo";
 import DangerZone from "./DangerZone/DangerZone";
 import { useAuth } from "../../context/LoginContext/LoginContext";
-import { userService } from "../../services/userService";
-import { paths } from "../../paths";
 import {
   dashboardViewContainer,
   dataFormsContainer,
 } from "./userDashboard.style";
 
 const UserDashboard = () => {
-  const navigate = useNavigate();
   const { userData } = useAuth();
   const { formData, updateUserProfile, isUsernameEditable } = useUserProfile();
   const { scores } = useScores(userData?.uid);
-
-  const handleDeleteAccount = async () => {
-    if (!userData?.uid) return;
-
-    if (
-      window.confirm(
-        "Are you sure you want to delete your account? This action cannot be undone."
-      )
-    ) {
-      try {
-        await userService.deleteUser(userData.uid);
-        alert("Account deleted successfully.");
-        navigate(paths.LOGIN);
-      } catch (error) {
-        console.error("Error deleting account:", error);
-        alert("Failed to delete account.");
-      }
-    }
-  };
 
   return (
     <Box sx={dashboardViewContainer}>
@@ -58,7 +35,7 @@ const UserDashboard = () => {
           </Grid>
           <Grid item sm={12} md={4}>
             <UserDataInfo />
-            <DangerZone handleDeleteAccount={handleDeleteAccount} />
+            <DangerZone />
           </Grid>
         </Grid>
       </Container>
